requirements.txt
# SPDX-FileCopyrightText: 2025 CERN (home.cern)
#
# SPDX-License-Identifier: BSD-3-Clause

pydantic==2.11.1
PyYAML==6.0.2
email_validator==2.2.0
requests==2.32.3
reuse==5.0.2
wemake-python-styleguide==1.0.0
yamllint==1.37.0
pytest==8.3.5
pytest-mock==3.14.0
